Test authentication for api
